Switch to pgzip
This is the actual package used in https://github.com/mholt/archiver
and positions itself as a drop-in replacement for compress/gzip, but
doing things in parallel.

Default concurrency is runtime.GOMAXPROCS(0) and the benchmarks can
be found here: https://github.com/klauspost/pgzip#performance

From pgzip's README.md:

>You should only use this if you are (de)compressing big amounts of data,
>say more than 1MB at the time, otherwise you will not see any benefit,
>and it will likely be faster to use the internal gzip library or this
>package.

Assuming there will be only a single compression job per task and most
caches are larger than 1MB in size, this should be fine to keep this
the default.

Resolves #50.