Introduce `CIRRUS_AGENT_EXPOSE_SCRIPTS_OUTPUTS`
In order to allow the agent to proxy all outputs of all scripts to the stdout/stderr. This is useful for folks having their own log collectors, for example, Stackdriver integration for GKE.

Made it opt-in since in this scenarios people pay for amount of logs collected, and we don't want to bring surprises to the bills.